worker/requirements.txt
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
boto3==1.29.7
reportlab==4.0.7
Pillow==10.1.0
python-dotenv==1.0.0


